+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Support for `scrollViewWillEndDragging`
 - Support for item's `copy` and `paste` action via `ActionPerformable`
 ### Changed
 - Deprecated `Item` in favour of `TableItem`